+++ src/net/neczpal/interstellarwar/common/game/SpaceShip.java
@@ -5,27 +5,27 @@ import java.io.Serializable;
 public class SpaceShip implements Serializable {
 	public static final int SPACESHIP_TYPES = 11;
 	private static final long serialVersionUID = 3683452581122892189L;
-	private static final int SPACE_SHIP_SPEED = 6;
+    private static final int SPACE_SHIP_SPEED = 6;
 
-	private int mUnitsNumber;
-	private int mOwnedBy;
-	private int mMaxTick;
-	private int mCurrentTick;
+    private int mUnitsNumber;
+    private int mOwnedBy;
+    private int mMaxTick;
+    private int mCurrentTick;
 
-	private Planet mFromPlanet;
-	private Planet mToPlanet;
+    private Planet mFromPlanet;
+    private Planet mToPlanet;
 
-	private float mVx, mVy;
+    private double mVx, mVy;
 
-	private int mTextureIndex;
+    private int mTextureIndex;
 
-	/**
-	 * Erstellt ein Raumschiff
-	 *
-	 * @param fromPlanet  Der Planet woher das Raumschiff beginnt
-	 * @param toPlanet    Der Planet wohin das Raumschiff einfahrt
-	 * @param unitsNumber Die Anzahl der Einheits, die transportiert wird
-	 */
+    /**
+     * Erstellt ein Raumschiff
+     *
+     * @param fromPlanet  Der Planet woher das Raumschiff beginnt
+     * @param toPlanet    Der Planet wohin das Raumschiff einfahrt
+     * @param unitsNumber Die Anzahl der Einheits, die transportiert wird
+     */
 	SpaceShip (Planet fromPlanet, Planet toPlanet, int unitsNumber) {
 		mTextureIndex = (int) (Math.random () * SPACESHIP_TYPES);
 
@@ -48,27 +48,27 @@ public class SpaceShip implements Serializable {
 
 	/**
 	 * Erstellt ein Raumschiff
-	 *
-	 * @param fromPlanet   Der Planet woher das Raumschiff beginnt
-	 * @param toPlanet     Der Planet wohin das Raumschiff einfahrt
-	 * @param vx           Die Geschwindigkeit von dem Raumshiff auf dem X-Achse
-	 * @param vy           Die Geschwindigkeit von dem Raumshiff auf dem Y-Achse
-	 * @param ownedBy      Die Nummer von Spieler, die diesem Raumschiff dominiert
-	 * @param unitsNumber  Die Anzahl der Einheits, die transportiert wird
-	 * @param currentTick  Die akutelle Zeitvariable
-	 * @param maxTick      Die maximalle Zeitvariable
-	 * @param textureIndex Das Index der Texture
-	 */
-	SpaceShip (Planet fromPlanet, Planet toPlanet, float vx, float vy, int ownedBy, int unitsNumber, int currentTick, int maxTick, int textureIndex) {
-		mTextureIndex = textureIndex;
-		mUnitsNumber = unitsNumber;
-		mCurrentTick = currentTick;
-		mMaxTick = maxTick;
-		mFromPlanet = fromPlanet;
-		mToPlanet = toPlanet;
-		mVx = vx;
-		mVy = vy;
-		mOwnedBy = ownedBy;
+     *
+     * @param fromPlanet   Der Planet woher das Raumschiff beginnt
+     * @param toPlanet     Der Planet wohin das Raumschiff einfahrt
+     * @param vx           Die Geschwindigkeit von dem Raumshiff auf dem X-Achse
+     * @param vy           Die Geschwindigkeit von dem Raumshiff auf dem Y-Achse
+     * @param ownedBy      Die Nummer von Spieler, die diesem Raumschiff dominiert
+     * @param unitsNumber  Die Anzahl der Einheits, die transportiert wird
+     * @param currentTick  Die akutelle Zeitvariable
+     * @param maxTick      Die maximalle Zeitvariable
+     * @param textureIndex Das Index der Texture
+     */
+    SpaceShip (Planet fromPlanet, Planet toPlanet, double vx, double vy, int ownedBy, int unitsNumber, int currentTick, int maxTick, int textureIndex) {
+        mTextureIndex = textureIndex;
+        mUnitsNumber = unitsNumber;
+        mCurrentTick = currentTick;
+        mMaxTick = maxTick;
+        mFromPlanet = fromPlanet;
+        mToPlanet = toPlanet;
+        mVx = vx;
+        mVy = vy;
+        mOwnedBy = ownedBy;
 	}
 
 	/**
@@ -89,38 +89,38 @@ public class SpaceShip implements Serializable {
 
 	public int getUnitsNumber () {
 		return mUnitsNumber;
-	}
+    }
 
-	public int getOwnedBy () {
-		return mOwnedBy;
-	}
+    public int getOwnedBy () {
+        return mOwnedBy;
+    }
 
-	public int getCurrentTick () {
-		return mCurrentTick;
-	}
+    public int getCurrentTick () {
+        return mCurrentTick;
+    }
 
-	public Planet getFromPlanet () {
-		return mFromPlanet;
-	}
+    public Planet getFromPlanet () {
+        return mFromPlanet;
+    }
 
-	public Planet getToPlanet () {
-		return mToPlanet;
-	}
+    public Planet getToPlanet () {
+        return mToPlanet;
+    }
 
-	public float getVx () {
-		return mVx;
-	}
+    public double getVx () {
+        return mVx;
+    }
 
-	public float getVy () {
-		return mVy;
-	}
+    public double getVy () {
+        return mVy;
+    }
 
-	public int getMaxTick () {
-		return mMaxTick;
-	}
+    public int getMaxTick () {
+        return mMaxTick;
+    }
 
-	public int getTextureIndex () {
-		return mTextureIndex;
+    public int getTextureIndex () {
+        return mTextureIndex;
 	}
 
 }
